% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle.R
\name{select_repeat_cols}
\alias{select_repeat_cols}
\title{Triple and shuffle columns in a dataframe with an even number of columns.}
\usage{
select_repeat_cols(df, offset, each = 3)
}
\arguments{
\item{df}{(dataframe) The dataframe to shuffle.}

\item{offset}{(integer) Whether to grab odd (1) or even (2) columns.}

\item{each}{(integer) How many times to repeat each column (default 3).}
}
\value{
Selected and tripled columns.
}
\description{
Given a dataframe with columns ABCD, return a dataframe
with columns AAACCC if offset is 1 or columns BBBDDD if offset is 2.
The new columns are given names x1..xN.
}
