% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle.R
\name{shuffle_data}
\alias{shuffle_data}
\title{Shuffle a dataframe, repeating columns.}
\usage{
shuffle_data(df, each = 3)
}
\arguments{
\item{df}{(dataframe) The dataframe to shuffle.}

\item{each}{(integer) How many times to repeat each column.}
}
\value{
The shuffled/duplicated dataframe.
}
\description{
Given a dataframe with columns ABCD and an even number of rows,
shuffle and duplicate the columns to produce (e.g.) AAACCCBBBDDD and shuffle
the rows from 1234 to 1324.
}
