% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle.R
\name{shuffle_indices}
\alias{shuffle_indices}
\title{Shuffle indices so that c(1, 2, 3, 4) maps to c(1, 3, 2, 4).}
\usage{
shuffle_indices(n)
}
\arguments{
\item{n}{(integer) The (even) number to shuffle.}
}
\value{
Shuffled indices.
}
\description{
Given an even number n, produce the vector c(1, 3, 2, 4, ...)
up to n.
}
